feat: 특정 구역 확대 로직 지도 컴포넌트 내부로 이동
